Logger

Create an object for logging database operations

Value

A new instance of the Logger

R6 class.

Public fields

log_path

(character(1))
A directory where log file is written (if this is not NULL). Defaults to getOption("SCDB.log_path").

log_tbl

The DB table used for logging. Class is connection-specific, but inherits from tbl_dbi.

start_time

(POSIXct(1))
The time at which data processing was started.

output_to_console

(logical(1))
Should the Logger output to console? This can always be overridden by Logger$log_info(..., output_to_console = FALSE).

Active bindings

log_filename

character(1)
The filename (basename) of the file that the Logger instance will output to

log_realpath

character(1)
The full path to the logger's log file.

Methods

Public methods

Logger$new()

Logger$log_info()

Logger$log_warn()

Logger$log_error()

Logger$log_to_db()

Logger$clone()

Method new()

Create a new Logger object

Usage

Logger$new(
  db_tablestring = NULL,
  log_table_id = getOption("SCDB.log_table_id"),
  log_conn = NULL,
  log_path = getOption("SCDB.log_path"),
  output_to_console = TRUE,
  warn = TRUE,
  ts = NULL,
  start_time = Sys.time()
)

Arguments

db_tablestring

A string specifying the table being updated

log_table_id

A DBI::Id() object or a character string readable by id, specifying the location of the log table.

log_conn

A database connection inheriting from DBIConnection

log_path

The path where logs are stored. If NULL, no file logs are created.

output_to_console

Should the Logger output to console (TRUE/FALSE)?

warn

Show a warning if neither log_table_id or log_path could be determined

ts

A timestamp describing the data being processed (not the current time)

start_time

The time at which data processing was started (defaults to Sys.time())

Method log_info()

Write a line to log file

Usage

Logger$log_info(
  ...,
  tic = Sys.time(),
  output_to_console = self$output_to_console,
  log_type = "INFO"
)

Arguments

...

One or more character strings to be concatenated

tic

The timestamp used by the log entry (default Sys.time())

output_to_console

Should the line be written to console?

log_type

A character string which describes the severity of the log message

Method log_warn()

Write a warning to log file and generate warning.

Usage

Logger$log_warn(..., log_type = "WARNING")

Arguments

...

One or more character strings to be concatenated

log_type

A character string which describes the severity of the log message

Method log_error()

Write an error to log file and stop execution

Usage

Logger$log_error(..., log_type = "ERROR")

Arguments

...

One or more character strings to be concatenated

log_type

A character string which describes the severity of the log message

Method log_to_db()

Write or update log table

Usage

Logger$log_to_db(...)

Arguments

...

Name-value pairs with which to update the log table

Method clone()

The objects of this class are cloneable with this method.

Usage

Logger$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

logger <- Logger$new(db_tablestring = "test.table",
                     ts = "2020-01-01 09:00:00")
#> Warning: log_path and log_table_id are both NULL and therefore NO LOGGING WILL BE DONE.
#> Consider adding options SCDB.log_table_id and/or SCDB.log_path to your .Rprofile