Gets a named table from a given schema

Gets a named table from a given schema

Usage

get_table(conn, db_table_id = NULL, slice_ts = NA, include_slice_info = FALSE)

Arguments

conn

An object that inherits from DBIConnection (as generated by get_connection())'

db_table_id

A DBI::Id() object or a character string readable by id. If missing, a list of available tables is printed.

slice_ts

If set different from NA (default), the returned data looks as on the given date. If set as NULL, all data is returned

include_slice_info

Default FALSE. If set TRUE, the history columns "checksum", "from_ts", "until_ts" are returned also

Value

A "lazy" dataframe (tbl_lazy) generated using dbplyr

Examples

conn <- get_connection(drv = RSQLite::SQLite())
#> Warning: No tables found. Check user permissions / database configuration:
#>   host: NULL
#>   port: NULL
#>   dbname: 
#>   user: NULL
#>   password: NULL

dplyr::copy_to(conn, mtcars, name = "mtcars")

get_table(conn)
#> [1] "Select one the following tables:"
#>   schema        table
#> 1   <NA>       mtcars
#> 2   <NA> sqlite_stat1
#> 3   <NA> sqlite_stat4
get_table(conn, "mtcars")
#> # Source:   table<`mtcars`> [?? x 11]
#> # Database: sqlite 3.41.2 []
#>      mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb
#>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1  21       6  160    110  3.9   2.62  16.5     0     1     4     4
#>  2  21       6  160    110  3.9   2.88  17.0     0     1     4     4
#>  3  22.8     4  108     93  3.85  2.32  18.6     1     1     4     1
#>  4  21.4     6  258    110  3.08  3.22  19.4     1     0     3     1
#>  5  18.7     8  360    175  3.15  3.44  17.0     0     0     3     2
#>  6  18.1     6  225    105  2.76  3.46  20.2     1     0     3     1
#>  7  14.3     8  360    245  3.21  3.57  15.8     0     0     3     4
#>  8  24.4     4  147.    62  3.69  3.19  20       1     0     4     2
#>  9  22.8     4  141.    95  3.92  3.15  22.9     1     0     4     2
#> 10  19.2     6  168.   123  3.92  3.44  18.3     1     0     4     4
#> # ℹ more rows

close_connection(conn)